Refactor usages of moved helper classes
