+++ tests/integration/test_pr_comment.py
@@ -31,9 +31,9 @@ from packit_service.models import (
 )
 from packit_service.service.db_triggers import AddPullRequestDbTrigger
 from packit_service.worker.allowlist import Allowlist
-from packit_service.worker.build import copr_build
-from packit_service.worker.build.copr_build import CoprBuildJobHelper
-from packit_service.worker.build.koji_build import KojiBuildJobHelper
+from packit_service.worker.helpers.build import copr_build
+from packit_service.worker.helpers.build.copr_build import CoprBuildJobHelper
+from packit_service.worker.helpers.build.koji_build import KojiBuildJobHelper
 from packit_service.worker.events.event import AbstractForgeIndependentEvent
 from packit_service.worker.jobs import SteveJobs, get_packit_commands_from_comment
 from packit_service.worker.monitoring import Pushgateway
@@ -45,7 +45,7 @@ from packit_service.worker.tasks import (
     run_koji_build_handler,
     run_testing_farm_handler,
 )
-from packit_service.worker.testing_farm import TestingFarmJobHelper
+from packit_service.worker.helpers.testing_farm import TestingFarmJobHelper
 from tests.spellbook import DATA_DIR, first_dict_value, get_parameters_from_results
